2019-06-18 by chisj

+++ src/main/java/org/study/heat/pojo/Payment.java
@@ -48,7 +48,7 @@ public class Payment {
     /**
      * 折扣比例
      */
-    private Float discount;
+    private Integer discount;
 
     /**
      * 实际费用
